// src/app/pages/dashboard/views/analyze/thumbnail-style/thumbnail-config-form/thumbnail.enums.ts
export enum FontWeight {
  BOLD = 'bold',
  EXTRA_BOLD = 'extra-bold',
}

export enum LogoSize {
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
}

export enum LogoPosition {
  TOP_LEFT = 'top-left',
  TOP_RIGHT = 'top-right',
  BOTTOM_LEFT = 'bottom-left',
  BOTTOM_RIGHT = 'bottom-right',
}

export enum Mood {
  CINEMATIC = 'cinematic',
  SERIOUS = 'serious',
  DRAMATIC = 'dramatic',
}

export enum DividerStyle {
  ARROW = 'arrow',
  LINE = 'line',
  VS = 'vs',
}